+++ test/alfrescoApi.spec.js
@@ -1,15 +1,9 @@
-/*global describe, it, beforeEach */
+/*global describe, it */
 
 var AlfrescoApi = require('../main');
 var expect = require('chai').expect;
 
 describe('Basic configuration test', function () {
-    beforeEach(function () {
-        this.host = 'http://127.0.0.1:8080';
-        this.contextRoot = 'script';
-        this.servicePath = 'alfresco';
-        this.scriptPath = 'testWebScript';
-    });
 
     it('default value', function () {
         var config = {
@@ -23,7 +17,7 @@ describe('Basic configuration test', function () {
 
     it('config parameter should be reflected in the client', function () {
         var config = {
-            host: 'http://testServer.com:1616',
+            hostEcm: 'http://testServer.com:1616',
             contextRoot: 'strangeContextRoot'
         };