+++ test/mockObjects/alfresco/nodeMock.js
@@ -268,6 +268,52 @@ class NodeMock extends BaseMock {
                 }
             });
     }
+
+    get200ResponseChildrenFutureNewPossibleValue() {
+        nock(this.host, {'encodedQueryParams': true})
+            .get('/alfresco/api/-default-/public/alfresco/versions/1/nodes/b4cff62a-664d-4d45-9302-98723eac1319/children')
+            .reply(200, {
+                'list': {
+                    'pagination': {
+                        'count': 2,
+                        'hasMoreItems': false,
+                        'totalItems': 2,
+                        'skipCount': 0,
+                        'maxItems': 100
+                    },
+                    'entries': [{
+                        'entry': {
+                            'createdAt': '2011-02-15T20:19:00.007+0000',
+                            'isFolder': true,
+                            'isFile': false,
+                            'createdByUser': {'id': 'mjackson', 'displayName': 'Mike Jackson'},
+                            'modifiedAt': '2011-02-15T20:19:00.007+0000',
+                            'modifiedByUser': {'id': 'mjackson', 'displayName': 'Mike Jackson'},
+                            'name': 'dataLists',
+                            'id': '64f69e69-f61e-42a3-8697-95eea1f2bda2',
+                            'nodeType': 'cm:folder',
+                            'parentId': 'b4cff62a-664d-4d45-9302-98723eac1319',
+                            'impossibleProperties' : 'impossibleRightValue'
+                        }
+                    }, {
+                        'entry': {
+                            'createdAt': '2011-02-15T22:04:54.290+0000',
+                            'isFolder': true,
+                            'isFile': false,
+                            'createdByUser': {'id': 'mjackson', 'displayName': 'Mike Jackson'},
+                            'modifiedAt': '2011-02-15T22:04:54.290+0000',
+                            'modifiedByUser': {'id': 'mjackson', 'displayName': 'Mike Jackson'},
+                            'name': 'discussions',
+                            'id': '059c5bc7-2d38-4dc5-96b8-d09cd3c69b4c',
+                            'nodeType': 'cm:folder',
+                            'parentId': 'b4cff62a-664d-4d45-9302-98723eac1319',
+                            'impossibleProperties' : 'impossibleRightValue'
+                        }
+                    }]
+                }
+            });
+
+    }
 }
 
 module.exports = NodeMock;